requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
Django==2.0
djangorestframework==3.7.3
idna==2.6
itypes==1.1.0
Jinja2==2.10
kombu==4.1.0
MarkupSafe==1.0
nameparser==0.5.4
psycopg2==2.7.3.2
pytz==2017.3
redis==2.10.6
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
